created mutations for events schema/typeDefs

+++ server/schemas/resolvers.js
@@ -3,13 +3,29 @@ const { signToken, AuthenticationError } = require('../utils/auth');
 
 const resolvers = {
   Query: {
-    profiles: async () => {
-      return Profile.find();
+    allprofiles: async (parent,args,context) => {
+      if (context.user) {
+        return Profile.find();
+      }
     },
 
-    profile: async (parent, { profileId }) => {
-      return Profile.findOne({ _id: profileId });
+    profile: async (parent, args, context) => {
+      if (context.user) {
+        return Profile.findOne({ _id: context.user._id });
+      }
     },
+    allevents: async (parent,args,context) => {
+      if (context.user) {
+        return Event.find();
+      }
+    },
+
+    event: async (parent,{eventId}, context) => {
+      if (context.user) {
+        return Event.findOne({ _id: eventId });
+      }
+    },
+
   },
 
   Mutation: {
@@ -35,11 +51,36 @@ const resolvers = {
       const token = signToken(profile);
       return { token, profile };
     },
-    removeProfile: async (parent, { profileId }) => {
-      return Profile.findOneAndDelete({ _id: profileId });
+    removeProfile: async (parent, { profileId }, context) => {
+      if (context.user) {
+        return Profile.findOneAndDelete({ _id: profileId });
+      }
     },
-    
-  },
+    addEvent: async (parent, { eventName, eventNotes, eventEnd, enableNotification, privacySettings }) => {
+      if (context.user) {
+        const event = await Event.create({ eventName, eventNotes, eventEnd, enableNotification, privacySettings });
+
+
+        return event;
+      }
+    },
+    addGuestList: async (parent, { eventID, guestList }) => {
+      if (context.user) {
+        const event = await Event.findOneAndUpdate({ _id: eventID }, { $push: { guestList } });
+
+
+        return event;
+      }
+    },
+    deleteEvent: async (parent, { eventId }, context) => {
+      if (context.user) {
+        return Event.findOneAndDelete({ _id: eventId });
+      }
+
+
+
+    },
+  }
 };
 
 module.exports = resolvers;
